Improving/Fixing crab3 utils, added dl script for latest TTree ntuple

+++ test/make_EdmNtuples.csh
@@ -10,16 +10,14 @@ set input_dir=$3
 set Nparallel=$4
 
 ls $input_dir | grep ".root" | awk '{ print "'$input_dir'/"$1 }' >! files.txt
+if ( -f make_edmtrees_"$version"_$name.csh ) mv make_edmtrees_"$version"_$name.csh make_edmtrees_"$version"_"$name"_prev.csh
 set n=`cat files.txt | wc -l`
-if ( -f make_edmtrees_"$version"_$name.csh ) then
-    mv make_edmtrees_"$version"_$name.csh make_edmtrees_"$version"_"$name"_prev.csh
-endif
 foreach i ( `seq 1 $n` )
     set infile=`sed -n "$i"p files.txt`
     set edmfile="$edmdir/B2GEDMNtuple_$i.root"
-    echo 'cmsRun ../../../../Analysis/B2GAnaFW/test/b2gedmntuples_cfg.py maxEvts=-1 sample="file:'$infile'" outputLabel="file:'$edmfile'" LHE=False' >>! make_edmtrees_"$version"_$name.csh
+    echo 'nice cmsRun $CMSSW_BASE/src/Analysis/B2GAnaFW/test/b2gedmntuples_cfg.py maxEvts=-1 sample="file:'$infile'" outputLabel="file:'$edmfile'" LHE=False' >>! make_edmtrees_"$version"_$name.csh
 end
 rm files.txt
 
 mkdir -p $edmdir
-source crab3/source_parallel.csh make_edmtrees_"$version"_$name.csh $Nparallel
+source $CMSSW_BASE/src/Analysis/B2GTTrees/test/crab3/source_parallel.csh make_edmtrees_"$version"_$name.csh $Nparallel
